feat: Enhance profile service with user-specific profile retrieval and update order service gRPC client configuration

+++ profile_service/src/main/java/com/lamnguyen/profile_service/domain/response/ProfileAdminResponse.java
@@ -21,6 +21,6 @@ import lombok.experimental.SuperBuilder;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProfileAdminResponse extends ProfileDto {
-	int totalOrders;
-	double totalSpent;
+	long totalOrders;
+	long totalSpent;
 }
